🐛fix: rebase 이후 ActivityService 중복 문제 해결
